Refactor all useIsDesktop and useIsMobile into ScreenFade #73
