+++ CHANGELOG.MD
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2020-05-17
+
+### Added
+
+- [Minor code refactoring](https://github.com/ragaeeb/das-js/issues/73)
+
 ## [0.2.0] - 2020-05-16
 
 ### Added
