+++ package.json
@@ -3,7 +3,7 @@
   "private": true,
   "homepage": "https://ragaeeb.github.io/das-js",
   "description": "Dar as-Sahābah website",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "license": "MIT",
   "scripts": {
     "build": "gatsby build",
@@ -20,8 +20,6 @@
     "adhan": "4.0.0",
     "bootstrap": "4.5.0",
     "chart.js": "2.9.3",
-    "cypress": "4.5.0",
-    "dotenv": "8.2.0",
     "gatsby": "2.21.33",
     "gatsby-cli": "2.12.21",
     "gatsby-image": "2.4.3",
@@ -51,6 +49,8 @@
   "devDependencies": {
     "@babel/plugin-transform-runtime": "7.9.6",
     "babel-eslint": "10.1.0",
+    "cypress": "4.5.0",
+    "dotenv": "8.2.0",
     "eslint": "7.0.0",
     "eslint-config-airbnb": "18.1.0",
     "eslint-config-prettier": "6.11.0",
