+++ src/components/About.jsx
@@ -1,8 +1,8 @@
-import Fade from '@kogk/react-reveal/Fade';
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useContext } from 'react';
 import { Container, Row } from 'react-bootstrap';
-import PortfolioContext from '../../context/context';
-import Title from '../Title/Title';
+import PortfolioContext from '../context/context';
+import ScreenFade from './ScreenFade';
+import Title from './Title';
 
 const renderService = ({ title, body }) => {
   return (
@@ -17,31 +17,18 @@ const About = () => {
   const { about } = useContext(PortfolioContext);
   const { paragraphOne, paragraphTwo, services } = about;
 
-  const [isDesktop, setIsDesktop] = useState(false);
-  const [isMobile, setIsMobile] = useState(false);
-
-  useEffect(() => {
-    if (window.innerWidth > 769) {
-      setIsDesktop(true);
-      setIsMobile(false);
-    } else {
-      setIsMobile(true);
-      setIsDesktop(false);
-    }
-  }, []);
-
   return (
     <section id="about">
       <Container>
         <Title title="About Us" />
         <Row className="about-wrapper">
-          <Fade left={isDesktop} bottom={isMobile} duration={1000} delay={1000} distance="30px">
+          <ScreenFade>
             <div className="about-wrapper__info">
               <p className="about-wrapper__info-text">{paragraphOne}</p>
               <p className="about-wrapper__info-text">{paragraphTwo}</p>
               <ul className="about-wrapper__info-text">{(services || []).map(renderService)}</ul>
             </div>
-          </Fade>
+          </ScreenFade>
         </Row>
       </Container>
       <br />
